Fix clippy cast_precision_loss and misc warnings
Fix cast_precision_loss warnings by avoiding precision loss:
- Break down time durations into hours/minutes/seconds components
- Convert station indices via i32 for lossless f64 conversion
- Use milliseconds for duration calculations to maintain precision

Fix other clippy warnings:
- Use f64::midpoint() instead of manual calculation
- Merge identical match arms in label positioning
- Use usize::from() for bool to int conversion
- Remove unused Props imports
- Extract helper functions to reduce nesting

Increase too-many-lines threshold to 150 for complex rendering code

+++ src/components/error_list.rs
@@ -37,7 +37,8 @@ fn ErrorListPopover(
                                         let conflict_message = conflict.format_message(station1_name, station2_name);
 
                                         let time_fraction = time_to_fraction(conflict.time);
-                                        let station_position = conflict.station1_idx as f64 + conflict.position;
+                                        // Convert via i32 to avoid precision loss (station counts are small)
+                                        let station_position = f64::from(conflict.station1_idx as i32) + conflict.position;
 
                                         view! {
                                             <div
